wrapper: force install, and running from install
The wrapper received 2 documented (README) CLI parameters,
so that using the wrapper becomes more convenient.

+++ SpecSAT.sh
@@ -8,6 +8,7 @@
 
 SCRIPT_DIR="$(cd -- "$(dirname -- "$0")" &>/dev/null && pwd)"
 declare -r SCRIPT_DIR
+FORCE_INSTALL="false"
 
 # Install error handler, which will act like 'set -e', but print a message
 error_handler() {
@@ -35,12 +36,17 @@ execute_silently() {
     fi
 }
 
+# Hidden command line options
 RUN_DIR="$PWD"
 if [ "${1:-}" = "--run-from-install" ]; then
     RUN_DIR="$SCRIPT_DIR"
     shift
 fi
 
+if [ "${1:-}" = "--force-install" ]; then
+    FORCE_INSTALL="true"
+    shift
+fi
 
 VENV_DIR="$RUN_DIR"/.SpecSATvenv
 declare -r VENV_DIR
@@ -48,13 +54,20 @@ declare -r VENV_DIR
 # Create virtual environment once
 if [ ! -d "$VENV_DIR" ]; then
     execute_silently python3 -m venv "$VENV_DIR"
+    # Activate virtual environment
+    source "$VENV_DIR"/bin/activate
+
+    # Install dependencies (for user)
+    execute_silently python3 -m pip install --upgrade -U -r "$SCRIPT_DIR"/requirements.txt
+else
+    # Activate virtual environment
+    source "$VENV_DIR"/bin/activate
 fi
 
-# Activate virtual environment
-source "$VENV_DIR"/bin/activate
-
-# Install dependencies (for user)
-execute_silently python3 -m pip install --upgrade -U -r "$SCRIPT_DIR"/requirements.txt
+if [ "$FORCE_INSTALL" = "true" ]; then
+    # Install dependencies (for user)
+    execute_silently python3 -m pip install --upgrade -U -r "$SCRIPT_DIR"/requirements.txt
+fi
 
 # Actually execute SpecSAT
 declare -i STATUS=0
